refactor: fe - objectFromServer 유틸 파일 분리 #111
