+++ frontend/src/pages/workspace/whiteboard-canvas/useSocket.ts
@@ -5,10 +5,7 @@ import { ClientToServerEvents, ServerToClientEvents } from '@pages/workspace/whi
 import { useRecoilState } from 'recoil';
 import { membersState } from '@context/workspace';
 import { fabric } from 'fabric';
-
-import cursorSvg from '@assets/icon/cursor.svg';
-import { v4 } from 'uuid';
-import { createCursorObject, moveCursorFromServer } from '@utils/fabric.utils';
+import { createCursorObject, moveCursorFromServer } from '@utils/object-from-server';
 
 function useSocket(canvas: React.MutableRefObject<fabric.Canvas | null>) {
 	const [members, setMembers] = useRecoilState(membersState);
